refactor(Join): split into more modules to avoid circular dependencies
